+++ composeApp/src/commonMain/kotlin/plus/vplan/app/data/repository/HomeworkRepositoryImpl.kt
@@ -159,19 +159,38 @@ class HomeworkRepositoryImpl(
             safeRequest(
                 onError = { return@channelFlow send(CacheState.Error(id.toString(), it)) }
             ) {
-                val metadataResponse = httpClient.get("${api.url}/api/v2.2/homework/$id")
-                if (metadataResponse.status == HttpStatusCode.NotFound && metadataResponse.isResponseFromBackend()) {
-                    vppDatabase.homeworkDao.deleteById(listOf(id))
-                    return@channelFlow send(CacheState.NotExisting(id.toString()))
+                val existing = vppDatabase.homeworkDao.getById(id).first()
+                var schoolApiAccess: SchoolApiAccess? = null
+                if (existing != null) {
+                    if (existing.homework.createdBy != null) schoolApiAccess = (vppDatabase.vppIdDao.getById(existing.homework.createdBy).first()?.toModel() as? VppId.Active)?.buildSchoolApiAccess()
+                    schoolApiAccess =
+                        if (schoolApiAccess != null) schoolApiAccess
+                        else (if (existing.homework.subjectInstanceId != null) vppDatabase.subjectInstanceDao.getById(existing.homework.subjectInstanceId).first()?.groups
+                            ?.mapNotNull { vppDatabase.groupDao.getById(it.groupId).first() }
+                            ?.mapNotNull { vppDatabase.schoolDao.findById(it.school.schoolId).first()?.toModel()?.getSchoolApiAccess() }
+                        else if (existing.homework.groupId != null) existing.homework.groupId.let {
+                            vppDatabase.groupDao.getById(existing.homework.groupId).first()?.school?.schoolId?.let {
+                                listOf(vppDatabase.schoolDao.findById(it).first()?.toModel()?.getSchoolApiAccess())
+                            }
+                        }
+                    else null)
+                            .orEmpty()
+                            .firstOrNull()
                 }
-                if (metadataResponse.status != HttpStatusCode.OK) return@channelFlow send(CacheState.Error(id.toString(), metadataResponse.toErrorResponse<Homework>()))
+                if (schoolApiAccess == null) {
+                    val metadataResponse = httpClient.get("${api.url}/api/v2.2/homework/$id")
+                    if (metadataResponse.status == HttpStatusCode.NotFound && metadataResponse.isResponseFromBackend()) {
+                        vppDatabase.homeworkDao.deleteById(listOf(id))
+                        return@channelFlow send(CacheState.NotExisting(id.toString()))
+                    }
+                    if (metadataResponse.status != HttpStatusCode.OK) return@channelFlow send(CacheState.Error(id.toString(), metadataResponse.toErrorResponse<Homework>()))
 
-                val metadataResponseData = ResponseDataWrapper.fromJson<HomeworkMetadataResponse>(metadataResponse.bodyAsText())
-                    ?: return@channelFlow send(CacheState.Error(id.toString(), Response.Error.ParsingError(metadataResponse.bodyAsText())))
+                    val metadataResponseData = ResponseDataWrapper.fromJson<HomeworkMetadataResponse>(metadataResponse.bodyAsText())
+                        ?: return@channelFlow send(CacheState.Error(id.toString(), Response.Error.ParsingError(metadataResponse.bodyAsText())))
 
-                val vppId = vppDatabase.vppIdDao.getById(metadataResponseData.createdBy).first()?.toModel() as? VppId.Active
-                val schools = vppDatabase.schoolDao.getAll().first().filter { it.school.id in metadataResponseData.schoolIds }.map { it.toModel() }
-                val school = schools.first()
+                    val schools = vppDatabase.schoolDao.getAll().first().filter { it.school.id in metadataResponseData.schoolIds }.map { it.toModel() }
+                    schoolApiAccess = schools.mapNotNull { it.getSchoolApiAccess() }.firstOrNull()
+                }
 
                 val homeworkResponse = httpClient.get {
                     url {
@@ -181,9 +200,9 @@ class HomeworkRepositoryImpl(
                         pathSegments = listOf("api", "v2.2", "homework", id.toString())
                         parameter("include_tasks", "true")
                     }
-                    vppId?.let { bearerAuth(it.accessToken) } ?: school.getSchoolApiAccess()?.authentication(this)
+                    schoolApiAccess?.authentication(this)
                 }
-                if (homeworkResponse.status != HttpStatusCode.OK) return@channelFlow send(CacheState.Error(id.toString(), metadataResponse.toErrorResponse<Homework>()))
+                if (homeworkResponse.status != HttpStatusCode.OK) return@channelFlow send(CacheState.Error(id.toString(), homeworkResponse.toErrorResponse<Homework>()))
                 val data = ResponseDataWrapper.fromJson<HomeworkGetResponse>(homeworkResponse.bodyAsText())
                     ?: return@channelFlow send(CacheState.Error(id.toString(), Response.Error.ParsingError(homeworkResponse.bodyAsText())))
 
@@ -209,10 +228,10 @@ class HomeworkRepositoryImpl(
                             cachedAt = Clock.System.now()
                         )
                     },
-                    homeworkTaskDoneAccount = if (vppId == null) emptyList() else data.tasks.map { it.value }.mapNotNull {
+                    homeworkTaskDoneAccount = if (schoolApiAccess !is SchoolApiAccess.VppIdAccess) emptyList() else data.tasks.map { it.value }.mapNotNull {
                         DbHomeworkTaskDoneAccount(
                             taskId = it.id,
-                            vppId = vppId.id,
+                            vppId = schoolApiAccess.id,
                             isDone = it.done ?: return@mapNotNull null
                         )
                     },
